adding more to tcBF dialog
